changed exceptions handling, changed work controllers.

+++ src/main/java/com/example/spring_boot_project/web/service/UserServiceImpl.java
@@ -1,5 +1,6 @@
 package com.example.spring_boot_project.web.service;
 
+import com.example.spring_boot_project.web.exeptions.UserNotFoundException;
 import com.example.spring_boot_project.web.model.User;
 import com.example.spring_boot_project.web.repository.UserRepository;
 import org.springframework.stereotype.Service;
@@ -16,6 +17,7 @@ public class UserServiceImpl implements UserService {
     public UserServiceImpl(UserRepository userRepository) {
         this.userRepository = userRepository;
     }
+
     @Override
     public void add(User user) {
         try {
@@ -30,29 +32,33 @@ public class UserServiceImpl implements UserService {
         try {
             Iterable<User> users = userRepository.findAll();
             List<User> result = new ArrayList<>();
-             users.forEach(result::add);
+            users.forEach(result::add);
             return result;
         } catch (Exception e) {
             e.printStackTrace();
         }
         return List.of();
     }
+
     @Override
-    public User getUser(long id) {
+    public User getUser(long id) throws UserNotFoundException {
         Optional<User> user = Optional.empty();
         try {
             user = userRepository.findById(id);
         } catch (Exception e) {
             e.printStackTrace();
         }
-        return user.orElse(new User()) ;
+        return user.orElseThrow(() -> new UserNotFoundException("Пользователь с таким id = " + id + ", не найден"));
     }
+
     @Override
-    public void removeUser(long id) {
-        try {
-           userRepository.deleteById(id);
-        } catch (Exception e) {
-            e.printStackTrace();
+    public void removeUser(long id) throws UserNotFoundException {
+
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+        } else {
+            throw new UserNotFoundException(
+                    "В процессе удаления возникла ошибка: пользователь с таким id = " + id + ", не найден");
         }
     }
 }
